requirements.txt
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
dash>=2.14.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
scikit-learn>=1.3.0
ta>=0.10.0
schedule>=1.2.0 